Add and configure Jinja2
It's necessary to retain the Django template processor so
django-compressor continues to work. What's here is the minimal config
required - APP_DIRS must be true as compressor is an app.

We can lean on the APP_DIRS behaviour of the Jinja2 templating backend
too, as long as we rename the templates/ folder to jinja2/. This seems
nice and unambiguous so I've done that. I've also renamed the files to
jinja files so syntax highlighting works properly in them.

Note that we have to disable bandit on the jinja Environment setup call
(see comment) so we'll do our own assertion that autoescape is True