Use ENV to determine which settings file to load
In practice manage.py only exists to set the DJANGO_SETTINGS_MODULE env
var before running `django-admin`. If we simply load .env at the top of
that file, we can have sane loading of different settings files
depending on environment.

Note that it's still necessary to load .env in base.py because it might
be invoked from another place (e.g. pytest, or gunicorn when we're in
production)