Build assets up front in the Dockerfile
Misleadingly, we COPY the contents of a local dev folder into the
container, which can include assets. When the container builds on CI,
those assets are definitely not going to be present.

Make it impossible to make this mistake (for current asset folders!) by
adding them to the .dockerignore so they don't get copied over on COPY
and then put explicit asset build steps into the Dockerfile, so we can
guarantee their presence (the container doesn't need to build them on
startup, for example).