refactor: Migrate static pages from Jinja & Django to Astro frontend
I didn't exactly match the styles when migrating over. This is
partly so we can move faster because we're now using Tailwind
instead of css selectors so it's a fair bit of effort to make it
match perfectly. Also, these pages (other than the home page) are
not frequently visited by users at all and likely to be redesigned
in the future.

I have also chosen not to remove the original jinja2 templates
because there are a lot of django tests that rely on them existing.
It will be easier to delete these tests in bulk when the migration
is complete than tease out which ones are affected by these changes.