refactor: Rename `Theme` model to `SelectedTheme`
As part of Theme Sign Off, we are introducing `CandidateTheme`s
which are the list of AI generated themes that can be selected from
during theme sign off. Renaming our existing `Theme` model to
`SelectedTheme` therefore makes it clearer that these are the themes
that have been selected during / after theme sign off.