fix: Allow users to view quetions when they don't have dashboard access
We recently made a change in https://github.com/i-dot-ai/consult/pull/883
to make a request to /api/consultations/{consultation_id}/questions/ in
the evaluation app to get the questions, rather than from with the
consultation resource. This introduced a bug where users who do not have
access to the dashboard could not do evaluations.