feat: Paginate selected and candidate theme list
This is default DRF behaviour which is bypassed by defined a `list`
method so we can add it back in by using `paginate_queryset` and
`get_paginated_response`.

Usually, there shouldn't be too many themes that it exceeds the
default page size of 100. But it's good to have pagination
behaviour in place if that ever changes.