fix: Docker caching in CI
We're currently hitting 429 Too Many Requests errors when building
images because we're not authenticating and there is a pull rate
limit of 100 pulls every 6 hours.

We could try setting up Docker Hub credentials but this is not the
approach other projects have taken. Instead, we should fix our
docker caching so that we're pulling from Docker Hub less often.
This is the suggested fix from Cookie-Cutter:
https://github.com/i-dot-ai/i-dot-ai-cookiecutter/blob/main/templates/_blueprints/containers/Makefile

This should also have the benefit of faster builds too.