refactor: Use permissions classes in ConsultationViewSet
I think this makes the permissions more explicit and is more robust
because it applies to all new methods we define too (in case we
forget about permissions for those methods).

I noticed, in the tests, we were referring to `consultation_slug`
as an identifier for the consultation but I think this is outdated
so have switched out to use `consultation_pk` instead.