fix: Show all consulations on /support/consultations page
This page is part of the support console which is only accessible
to admin users. We want to list all the consultations here so that
admin users (members of i.AI) can easily add themselves and others
to consultations.

In order to do this, I have made the default (`get_queryset`)
behaviour return all consultations for admin users and only
assigned consultations for non-admin users. For the /consultations
page, we just want to display the consultations the user is a
member of, regardless of whether they are a admin or not. I have
therefore defined a new endpoint `/consultations/assigned` to
support this.