fix: Failing tests due to clocks going back
Interestingly, the failing tests were not directly because of a
timezone difference. It was because the iso strings in the API
response uses 'Z' offset, whereas Python's datetime `isoformat`
method uses '+00:00' to represent UTC. So, because the clocks
went back yesterday to GMT (which is the same as UTC), it caused
our tests to fail.