fix: Consultation questions not showing in dashboard
I recently made a change to the global `Routes` so that we would
be showing just assigned consultations on /consultations. I hadn't
realised, however, that this is used elsewhere and therefore the
additional '?scope=assigned' was messing up requests for the nested
questions resource used in the dashboard and elsewhere.