chore: Update dummy_data script to support Theme Sign Off
I have updated the script so that it can create a consulation at
either Theme Sign Off stage or Analysis stage. This should make it
easier to get the data we need for QA and development.

I deleted the test_managing_consultations_via_support test because
we removed the button on /support/consultations page to generate
dummy consultations so this test is no longer needed.