feat: Add search mode for getting representative repsonses for a theme
This implements the semantic search strategy implementation[1] put
forward by Alex and Michael with one difference, where we're using
`SearchRank` as an approximation for BM25.

I have also added tests for each of our search modes. Because we're
only calling the AzureOpenAI client to create embeddings in deployed
environments, I have defined pre-calculated embeddings for each of
the responses and search values in the test data.

[1] https://incubatorforartificialintelligence.atlassian.net/wiki/spaces/Consult/pages/32079924/Semantic+Search+Strategy+Recommendation