fix: Failed "Sign-off" (finding themes) jobs
Across all environments, we are getting the error:
  ModuleNotFoundError: No module named 'pydantic'

This is a regression caused by https://github.com/i-dot-ai/consult/pull/1019/changes
where we moved the `from pydantic import BaseModel` import to the
top of the file instead of where we had it before, after runtime
installation completed.

Instead of fixing this by moving the import back to where it was,
I have decided to move the installation step into the Docker build
stage. This ensures we have the modules we need. We did have it like
this originally but changed it as part of https://github.com/i-dot-ai/consult/pull/935
to ensure we always have the latest themefinder release. I think it's
better practice to pin the version of themefinder to ensure
reproducability and consistency. We can always update the version
whenever we need to.